bug fixes: rbac, body parser for visitor post lambda and cosmatic changes to code
