+++ docs/VMS_swagger.yaml
@@ -13,8 +13,6 @@ tags:
       description: Operations related to ID cards
     - name: Approval
       description: Operations related to approvals
-
-
 paths:
     /login:
         post:
@@ -131,7 +129,6 @@ paths:
                                         type: string
                                         description: status
                                         example: success
-                    
 
                 "401":
                     description: Unauthorized
@@ -417,7 +414,7 @@ paths:
                 httpMethod: POST
                 contentHandling: CONVERT_TO_TEXT
                 type: aws_proxy
-                
+
             security:
                 - CognitoUserPool: []
         put:
@@ -1519,7 +1516,6 @@ paths:
                 type: aws_proxy
             security:
                 - CognitoUserPool: []
-
 
 components:
     securitySchemes:
